+++ client/app/styles/colors.ts
@@ -1,6 +1,7 @@
 // BC GOV colors
 export const BC_GOV_LINKS_COLOR: string = "#1A5A96";
 export const BC_GOV_PRIMARY_BRAND_COLOR_BLUE: string = "#003366";
+export const BC_GOV_BACKGROUND_COLOR_BLUE: string = "#38598a";
 // App specific colors
 export const DARK_GREY_BG_COLOR: string = "#E5E5E5";
 export const LIGHT_GREY_BG_COLOR: string = "#fafafc";
