+++ client/app/user-operator/page.tsx
@@ -1,5 +1,5 @@
-import Loading from "@/app/components/loading";
 import { Suspense } from "react";
+import Loading from "@/app/components/loading/SkeletonGrid";
 import PageBar from "@/app/components/PageBar";
 import UserOperatorForm from "@/app/components/form/UserOperatorForm";
 import { userOperatorSchema } from "@/app/utils/jsonSchema/userOperator";
@@ -7,7 +7,7 @@ import { userOperatorSchema } from "@/app/utils/jsonSchema/userOperator";
 export default async function Page() {
   const pageBarLabel: JSX.Element = (
     <div>
-      <small>Select Operator</small> {">"} <b>Request Access</b>
+      <small>Select Operator {">"}</small> <b>Request Access</b>
     </div>
   );
   return (
